+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed an issue with solutions like `react-refresh` that might not work when using RRT without React Devtools
 - Fixed an issue where in some cases events were not loaded because the `getEvents()` remote method is not yet available
 - Fixed displaying changes in a context's Provider
+- Added "Waiting for React..." when no events have been received for some time since UI initialization
 
 ## 0.3.0 (September 8, 2021)